RAG/requirements.txt
# Core dependencies
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# MCP Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# IBM Watsonx.ai
ibm-watsonx-ai>=0.2.0
ibm-watson-machine-learning>=1.0.0

# Vector Store
pymilvus>=2.3.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.0

# Numpy (required by various packages, install first)
numpy>=1.24.0,<2.0.0